As a first step, use Tonelli-Shanks for cubic sqrt
